Accelerometer control implementation, and more
- Separated PlayerController into PlayerOriginController and PlayerObjectController, in order to separate positioning/other jobs
- Changed rotation calculation for player, make player rotate around origin
- Added acclerometer control to player
- Added adjustable accelerometer, with support of recalibration
- Added setting page with real time preview and options UI
- Added options to increase/decrease sensitivity of accelerometer
- Configured Unity to build Windows 10 apps

+++ Graphics Project 2/Assets/GameSettings.cs
@@ -1,4 +1,3 @@
 ﻿public class GameSettings  {
-    public int acclerometerSensitivity = 10;
-
+    public float acclerometerSensitivity = 10f;
 }
